+++ DBAL/Type/ImplodedArray.php
@@ -6,30 +6,9 @@ namespace Trikoder\Bundle\OAuth2Bundle\DBAL\Type;
 
 use Doctrine\DBAL\Platforms\AbstractPlatform;
 use InvalidArgumentException;
-use LogicException;
 
 trait ImplodedArray
 {
-
-    /**
-     * {@inheritdoc}
-     */
-    public function convertToDatabaseValue($value, AbstractPlatform $platform)
-    {
-        if (!\is_array($value)) {
-            throw new LogicException('This type can only be used in combination with arrays.');
-        }
-
-        if (0 === \count($value)) {
-            return null;
-        }
-
-        foreach ($value as $item) {
-            $this->assertValueCanBeImploded($item);
-        }
-
-        return implode(self::VALUE_DELIMITER, $value);
-    }
 
     /**
      * {@inheritdoc}
