+++ DBAL/Type/RedirectUriOdm.php
@@ -5,9 +5,11 @@ declare(strict_types=1);
 namespace Trikoder\Bundle\OAuth2Bundle\DBAL\Type;
 
 use Doctrine\ODM\MongoDB\Types\Type;
+use LogicException;
 use Trikoder\Bundle\OAuth2Bundle\Model\RedirectUri as RedirectUriModel;
 
 use function explode;
+use function implode;
 
 /**
  * Class RedirectUriOdm
@@ -36,6 +38,26 @@ final class RedirectUriOdm extends Type
         $values = explode(self::VALUE_DELIMITER, $value);
 
         return $this->convertDatabaseValues($values);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function convertToDatabaseValue($value)
+    {
+        if (!\is_array($value)) {
+            throw new LogicException('This type can only be used in combination with arrays.');
+        }
+
+        if (0 === \count($value)) {
+            return null;
+        }
+
+        foreach ($value as $item) {
+            $this->assertValueCanBeImploded($item);
+        }
+
+        return implode(self::VALUE_DELIMITER, $value);
     }
 
     /**
